Stop shadowing `ONE`
Our loosening of the overflow safety surface introduced a warning, which
is prominently visible on our [canonical godbolt link]:

![image](https://github.com/aurora-opensource/au/assets/1819744/d55ef99f-7073-4644-9ffc-f75fa3893441)

Oops!  We shadowed the _magnitude_ `ONE` (globally available in the
library) with the _constant_ `ONE`.

The quick fix is to rename the newcomer (which is not publicly visible
anyway) to `UNITY`.

[canonical godbolt link]: https://godbolt.org/z/KrvfhP4M3