Add utilities to check for overflow
Each magnitude applying type gets its own dedicated utility.  This is
nice because the overflow condition appears directly next to the
function that applies the magnitude, so it's easy to check.  For
example, when applying a rational magnitude to an integral type, we
check whether the _numerator alone_ would cause overflow, because we
know we'd be multiplying by that numerator before dividing by the
denominator.

We also make a new type trait to make it easy to get the appropriate
type of magnitude applicator.  This made it a lot easier to write the
tests.