Add utilities to check for truncation
Once again, each magnitude-applying type gets its own dedicated utility,
bringing the same benefits as for the overflow checker.  When it comes
to truncation, most helpers get a much simpler implementation: it's
simply always `false` if we're doing a pure integer multiply, or if the
type itself is floating point.  On the _implementation_, we do get an
extra level of indirection, because the pure-integer-divide application
results are different depending on whether or not our type is floating
point.