Handle numerators that fit P, but not T
This is an edge case that I had missed before.

The rational magnitude applier helper libraries all check whether `N`
and `D` can fit in the _promoted_ type of `T`, not `T` itself.  This is
what they _should_ do, because this is the true limit of what's possible
in the computation.  However, on the _implementation_ side, we were
getting the value in the type `T`.  I was able to write a test case that
exposed this, and now the test case passes.

We should not need to worry about this for the other `ApplyMagnitude`
specializations, because the `x * N / D` is the only case we have where
the limits of the promoted type can produce a different outcome than the
limits of the target type.