Migrate to new home for LLVM toolchain (#225)
The LLVM toolchain graduated from `grailbio/bazel-toolchain` to
`bazel-contrib/toolchains_llvm`.  Good for them!  Unfortunately, that
broke everyone's build who was using a `sha256` checksum, because GitHub
does not guarantee SHA stability for automatically generated artifacts.
(Remember when they [broke the world] in January 2023?)

We need to upgrade to the new repo location, which means changing the
SHA.  We would not like to be forced to do this again, so we'll want to
link to a concrete artifact, not an auto-generated one.  This means
moving from a particular commit hash to a named release, because the
`.tar.gz` file is attached to that release.  So we pick the latest one.

I followed and adapted the [0.10.3 instructions], using the "not
`bzlmod`" section.  I also grepped the repo for `grailbio`, and updated
a URL in some comments.  (I verified that the new URL gives the same
contents as the old one.)

But wait, there's more!  The `aspect-build/gcc-toolchain` repo _also_
migrated to a new home!  Now it lives with a particular individual
called `f0rmiga`, instead of the `aspect-build` bazel organization.  We
can't land code until we unbreak _all_ of the builds, so we need to fix
that in the same PR.  Here, I _don't_ want to migrate to a new release,
because (a) we have a patch on top of the specific commit that we had
been using, and (b) they don't even upload archive files for their
releases.

[broke the world]:

https://github.blog/2023-02-21-update-on-the-future-stability-of-source-code-archives-and-hashes/
[0.10.3 instructions]:
https://github.com/bazel-contrib/toolchains_llvm/releases/tag/0.10.3

Fixes #224.