Support `QuantityPoint` in rounding functions (#237)
Includes the `round`, `floor`, and `ceil` families.

There's more subtlety here than for `Quantity`, because we can have
additive offsets.  However, since these functions all aim to behave like
the standard library, and the standard library always uses floating
point, then it's actually quite simple.  The way we handle rounding a
quantity point in different units than it's stored is to first perform
the conversion, then do the rounding.  As with quantities, we use
`RoundingRep` to make sure we're using a sane floating point type to
perform the conversion.

Includes a bunch of new test cases, and doc updates.

Fixes #221.