Deprecate `PI`
If we have anything called `PI` --- even if it is properly encapsulated
inside a C++ namespace! --- then users will be unable to use our library
if they also depend on _any library_ that defines a macro called `PI`,
in _any_ namespace (because macros _do not care_ about namespaces).

Of course, defining a macro named `PI` would be an execrable and
amateurish violation of long-accepted best practices for C++ software
engineering.  However, at least one library does this: the Arduino core
library.  Therefore, if we want to support it --- which we most
certainly do --- then we will have to yield.

The plan going forward is to release this deprecation as part of 0.3.5,
and then fully delete the variable as part of 0.4.0.  In the meantime,
users can still get Au to work; they just need to make sure that they
include Au _after_ any files that --- _lamentably_ --- define a macro
`PI`.

We have other "short all-caps" identifiers, `ONE` and `ZERO`.  We could
get rid of `ONE` with no trouble at all; `mag<1>()` is a fine
replacement.  Getting rid of `ZERO` would be extremely painful, though.
The motivation to define a `ZERO` macro is far less than for `PI`, so I
am planning to simply hope that no library of note has done this, and
expecting to tell users of any such library that they are simply out of
luck.

Fixes #246.