Set minimum acceptable version for find_package
I tested this out by actually installing googletest system-wide.  It
turned out to be version 1.11.  When I wiped the build folder and
re-installed, there was still a `_deps` folder that had googletest in
it.  Then when I set the new version arg to 1.10, the install was a lot
shorter and there was no `_deps`.