Prep for 0.3.5 release
I updated the CMake version number to 0.3.5.

Separately, in the docs, I fixed up our bazel installation instructions,
because they were referring to a very old release.  The major change
here was to point users towards the manually uploaded archive file,
rather than the automatically generated one, because the former is
stable but the latter is not.