Prep for 0.3.5 release (#271)
I updated the CMake version number to 0.3.5.  Note that the URL doesn't
exist yet.  This is expected (see `RELEASE.md`); we'll make it exist as
part of the release process.

Separately, in the docs, I fixed up our bazel installation instructions,
because they were referring to a very old release.  The major change
here was to point users towards the manually uploaded archive file,
rather than the automatically generated one, because the former is
stable but the latter is not.