Update to new Au logo
This brings us in line with Aurora's branding refresh.  These new assets
were designed directly by Ethan Bennett.

I tested by running `au-docs-serve` and checking that the favicon was
updated, and by pushing the branch remotely and viewing the rendered
README.  No other assets needed to be updated (`aurora-logo.png`, used
in our AVdocs, is white-on-transparent and was thus unaffected by the
change.)