Fetch gcc toolchain from `archive/` URL
Previously, we were using the `tarball/` URL, which includes both the
organization and the repo name as part of the prefix.  GitHub also has
an `archive/` URL format which, according to this [comment], only
includes the repo name, not the organization!

This PR switches to the new format, making us less vulnerable to build
breakage if the repo changes homes again.  (Notice how `f0rmiga` never
appears except in the URL --- in particular, it is now absent from
`strip_prefix`.)  The even better solution would be to download stable
tarballs from the release page, but the project has not yet begun to
provide those.

[comment]: https://github.com/f0rmiga/gcc-toolchain/issues/165#issuecomment-2290006934