Change compiler naming scheme for MSVC (#286)
We've been calling the version on Windows Server 2022, "MSVC x64 19.35".
Well, apparently it's now running 19.40, not 19.35.  If this can change
without our knowing, then putting the version number in the name is
probably not a great idea.

An alternative suggestion was to use the version of Windows Server in
the name, since that won't ever change.  We do that here.

This will break the continuity for the page giving historical results
per job, i.e., it will create a "new job".  I'm not especially concerned
about that.

Fixes #283.