Consolidate Au CMake targets
This breaks the 1:1 correspondence between target and test, so we split
up our helper accordingly into two.

We make `au` the single mega-target that holds everything that doesn't
depend on GTest.  It turns out that there are two targets that do.
`testing` is for external users, and contains custom GMock matchers.
`chrono_policy_validation` is an internal-only library that makes it
easy to test that Au is like chrono where we want to be, and unlike
chrono otherwise.  Each of these targets also depends on the `au`
mega-target.