Refresh installation instructions (#302)
The old flow of the page dates from the first open source release, when
only bazel was fully supported, and we provided the single-file
packaging as the "lowest common denominator" fallback that worked with
everybody.  This summer, we added official CMake support, and just last
week, the community also added support for both conan and vcpkg.  This
radically changes and simplifies the recommended installation approach:
now "full install" is the right choice for almost everyone! Accordingly,
it's time to refresh our instructions.

- Make up-front guidance simpler and more direct
- Rearrange the table: put Full Install first; unify Full Install
  columns; clean up and explain some text
- Remove decision flowchart: things are much simpler now, and we don't
  really need it
- Refresh the bazel instructions to mention 0.3.5 (might as well)
- Add links and instructions for installing with package managers
- Move the single-file section _after_ the full install one

I used "checkpoint commits" for this PR, so this last change is in a
separate "pure-move" commit within the PR, to make this easier to
review.

The reviewer can also run `au-docs-serve` and do A/B comparison between
two tabs:
- https://aurora-opensource.github.io/au/main/install/
- http://127.0.0.1:8000/au/install/

Fixes #301.