Test options in CMake CI jobs
Currently, we only run one CMake configuration in CI.  It's the default
one, where we permit the GTest dependency, and we build and run tests.
Now we are adding two more configurations, one for each of the new
options:

- Set `AU_ENABLE_TESTING` to `FALSE`
- Set `AU_EXCLUDE_GTEST_DEPENDENCY` to `TRUE` (implies the former)

In both these instances, the `test` target won't exist, so we exclude it
from running.

We do however retain the `CMAKE_VERIFY_INTERFACE_HEADER_SETS` option,
because that gives us useful testing for our interface targets.

Fixes #304.