Update troubleshooting guide
The biggest change is that we add a new section for the new kind of
error, when users try to pass something to a `QuantityMaker` even though
it's already a `Quantity` (and similar for `QuantityPoint`).

We also update the guidance to recommend `unblock_int_div` instead of
the obsolete `integer_quotient`.

Other than that, the PR basically consists of running the build under
the five configurations shown here, and copy-pasting error messages.
Most changes are just line number tweaks.  Some changed implementation
details do show up in the error messages.  Also, sometimes the compiler
changed to be more or less verbose.

The details of the changes aren't really important (assuming that I
copy-pasted correctly).  It's just an update.

Fixes #288.